+++ src/app/app.component.ts
@@ -11,7 +11,6 @@ import { Component } from '@angular/core'
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss'],
   animations: [
     trigger('routerAnimation', [
       // Move Right to Left <-----