+++ src/app/app.module.ts
@@ -10,6 +10,7 @@ import { FlexLayoutModule } from '@angular/flex-layout'
 import { FormsModule, ReactiveFormsModule } from '@angular/forms'
 import { MatButtonModule } from '@angular/material/button'
 import { MatDialogModule } from '@angular/material/dialog'
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner'
 import { MatSelectModule } from '@angular/material/select'
 import { MatSnackBarModule } from '@angular/material/snack-bar'
 
@@ -29,9 +30,15 @@ import { SharedModule } from '@shared/shared.module'
 import { Ng2ImgToolsModule } from 'ng2-img-tools'
 import { AppRoutingModule } from './app-routing.module'
 import { AppComponent } from './app.component'
+import { AboutComponent } from './pages/leaderboard/dialogs/about/about.component'
 
 @NgModule({
-  declarations: [AppComponent, LeaderboardComponent, LoginComponent],
+  declarations: [
+    AppComponent,
+    LeaderboardComponent,
+    LoginComponent,
+    AboutComponent
+  ],
   imports: [
     AngularFireAuthModule,
     AngularFireModule,
@@ -50,7 +57,8 @@ import { AppComponent } from './app.component'
     ScrollingModule,
     FlexLayoutModule,
     SharedModule,
-    MatSelectModule
+    MatSelectModule,
+    MatProgressSpinnerModule
   ],
   providers: [
     AuthService,