+++ src/app/pages/leaderboard/dialogs/about/about.component.ts
@@ -1,0 +1,16 @@
+import { Component } from '@angular/core'
+
+import { MatDialogRef } from '@angular/material'
+
+@Component({
+  selector: 'app-about',
+  templateUrl: './about.component.html',
+  styleUrls: ['./about.component.scss']
+})
+export class AboutComponent {
+  constructor(private dialogRef: MatDialogRef<AboutComponent>) {}
+
+  closeModal(): void {
+    this.dialogRef.close()
+  }
+}